Update the docs
This update adds several do-it-the-easy-way sections documenting
the usage of the new Makefile and clarifies other possibly unclear or
incomplete parts.  It also drops the final hashes / reflow section that
is no longer needed.  It extends the list of dependencies, too.

And finally it reflows the file with `fmt`, for which you will hate me.

Signed-off-by: Petr Šabata <contyk@redhat.com>