Dropping perl from the set
Via ppc64-utils, powerpc-utils and s390utils* removal from the toplevel
lists.  Just like with kernel-tools, this a temporary fix.  We expect to
support these tools in a module living on top of Platform as it appears
to be impossible to drop their dependency on Perl without a major
rewrite.

This update also drops a number of packages we do want in Platform but
they weren't in the toplevel list.  This will be fixed in a future
update, once the shared-userspace packages land.

Signed-off-by: Petr Šabata <contyk@redhat.com>